Add float

+++ calc2.py
@@ -19,7 +19,7 @@ print("***Kalkulator***")
 con = "t"
 while con == "t" or con == "T":
     action = int(input("Wybierz działanie:\n [1] - dodawanie\n [2] - odejmowanie\n [3] - mnożenie\n [4] - dzielenie\nTwój wybór: "))
-    liczba1 = int(input("Podaj pierwszą liczbę: "))
-    liczba2 = int(input("Podaj drugą liczbę: "))
+    liczba1 = float(input("Podaj pierwszą liczbę: "))
+    liczba2 = float(input("Podaj drugą liczbę: "))
     logging.info(f"Wynik: {calc(action,liczba1,liczba2)}")
     con = input("Kontynuować? [T] - tak, [N] - nie: ")
